Introduce HCPConfig as a general configuration interface
This interface will be used by the SDK as well as other HCP tools as a
common source of configuration values.

The default configuration will provide values to interact with the
production API of HCP. Custom configuration value can be provided by the
intergrating application and/or through default sources like environment
variables.

The sources will likely be extended in the future. E.g with the option
to read the configuration from a file.