transform-swagger: fix GoogleRPCStatus case
When updating cloud-packer-service to grpc-gateway v2, we broke the
generated SDK for the service because the error type for all our API
calls was changed from #/definitions/grpc.gateway.runtime.Error to
\#/definitions/google.rpc.Status.

This change then conflicted with the transform-swagger code, as it would
invoke `toCamel' on it, rendering the generated type as
`*cloud.GoogleRpcStatus', where the real type is
`*cloud.GoogleRPCStatus'.

To circumvent this problem, we hard-code this special case in the
transform command, so we end-up with the expected type in the generated
Go code.