Introduce a common cache used for all kinds of login
This will allow to cache user, service principal and workload identity
provider tokens in the same way. It will also provide a common way to
refresh and fetch new tokens if the cached token is expired.

This new cache will replace the user specific session and cache solution
that is currently used.