Use loginSourceType instead of login boolean flag
It was initially planned to support service-principal login, but since
that is not implemented right now the distinction between source types
can rely on the sourceType and the complexity can be spared. In the
future a flag might need to be added to differencate between user and sp
based login.